+++ docs/release-notes.md
@@ -5,6 +5,16 @@ hide:
 
 # Release notes
 
+## Version 0.4.0
+
+### Added
+
+### Changed
+
+- When string or class is provided by a callable for settings it is parsed and cached.
+  This allows lazy parsing of environment variables, so they can be changed programmatically.
+- Allow unsetting settings via with_settings by using False or "".
+
 ## Version 0.3.0
 
 ### Breaking